move `StrictCostDispatcher#construct` to Utils
